Added smooth scrolling, removed an old playback file, updated readme.
The big change here is the addition of smooth scrolling. In the previous
version moving from line to line worked but the change was abrupt. This
version uses JQuery's animate() function to make the transition less
jarring.

I also removed an old copy of the playback html file.

I updated the readme to include the options for sharing a playback (this
includes the relatively recent added 'download the playback.html' file
option)