Finished marking recent changes
Made highlighting recent changes an optional feature instead of a required one.

The initial state of a repo might be quite large so the initial starting step of a playback is not considered a recent change (since removing the recent change css is slow when there are 10k's of inital elements).

Added a change highlight in the name of the files (the tabs at the top of the playback window) that holds a recent change so that someone can see which files changed since the last recent change toggle.