Added comment markers
Added comment markers to the slider whenever a comment is present (these can be removed whenever 'pause for comments' is turned off).

Added buttons to move to the next comment.

Moved the event count out of the slider and into the event info area.

Added an alert area/method to notify of errors (use for when adding/deleting a comment on the server fails).

Added keyboard shortcuts for all of the playback control options.

Removed all references to glyphicons.