Title and description for a playback
Added an editable playback title and description that can be used to set the context for a playback.

Made the text of comments editable.

Changed how comment markers are displayed when they are turned off (no they are opaque so the user can still see where comments would be even if they are turned off).

Moved the comment portion of the screen to the left from the right to make it easier to trace comments and their selected code.

Made stopping for a comment optional when making large steps through a playback. This is mostly for when someone uses the slider to go anywhere in the playback the user won't be stopped for each comment.

Made filtering work with comments. The comment markers are at the correct scale now (they weren't before).

Optimized the way we check for initial text that is added when a non-empty project is started. This text is dumped in one big block and not counted as a recent change.