Copy and paste
Added cut/copy/paste support within a repo. During playback all pasted code will be identified. If the code was pasted from this repo its history can be traced back to its origins. The user must add key bindings in order for this to work or go through the palette to copy/cut/paste.

The selected text filter now allows other filters to be used in combination with it. For example, there might be some text selected and a time filter added so that you will only see the selected text that also satisfies the time filter.

The user can now select multiple pieces of code at the same time. This can be used for comments and for selected text filtering. We have our own highlighting for selected text. Multiple selections can be made with the control key pressed.

There is now an option in the filter modal to allow paste history to show up when filtering. By default this option is off.

We can move to the beginning or end of a playback using the ctrl-shift-arrow combinations. All comments will be ignored.

More information about every event is displayed during playback: the character just displayed and whether it was copied from somewhere else (local repo or some other source).