First attempt at version control
A storyteller project can now be merged with another one (if they have some common ancestory) and any line conflicts will be marked in a playback.

The user can view the two versions with a playback and if they like the merged results they can download the combined repo.

If there are line conflicts they will be highlighted in the downloaded project and can then be cleaned up in the editor.

Adding this feature required adding a branch id to all ids in the system. It is a five digit base 62 number.

I also changed the compression to be off by default.