Made the playback title/description show up if there is a user supplied description
Previous to this version the detailed title and description of a playback was hidden by default. Now, if there is a user supplied description the detailed title/description panel will be shown.

This ensures that if there is a description for a playback that the viewer will see it. Without this, the user would have to manually check to see if there was one.