Switched the default cut/paste playback behavior
Switched the default from showing all cuts/pastes in blocks to animating all cuts/pastes.

My thinking is that if people are watching the animation they really want  to see the code evolve. It can be a little jarring to move through a playback with a lot of cuts/pastes.

I also switched the order of the download options and selected to not go beyond the current point in the playback rather than the full undo.