Made comments (somewhat) editable
Comment editing:
- you can now edit the highlighted code for a comment in addition to the text of the comment
- you can now move comments to different positions in the comment list for a pause point

Some minor changes related to editing:
- can only edit a comment that is selected
- when a comment is selected and moved to edit mode the highlighted code becomes selected code
- made the selected code css have a higher precedence than 'comment relevant' code
- when code is both selected and 'comment relevant' there is a new css style

What this does not do:
- can't edit images in a comment.

My thinking is that this would require a lot of work. Instead, if a user wants to edit the images in a comment they can create a new one and add the new images to the comment.