Fixed bug preventing pasting a pic
In fixing the undo/redo with ctrl-z/ctrl-y in the drawing modal I added a e.preventDefault() whenever 'ctrl' happens. In other words, I consumed the 'ctrl' event no matter what.

I want to consume the event only when I know it is a ctrl-z or a ctrl-y. So, I move the e.preventDefault() inside the 'if' only when I know these combinations happen.