Turn on Storyteller instead of default on
The main change in this commit is that Storyteller is not automatically turned on when the plugin is downloaded. Now, the user has to actively start it if they want to start tracking changes. The user gets notified when Storyteller is active on startup of a workspace and if it is not active in a workspace. If they attempt to start without having a workspace open they will get a notification.

Some other changes:
- Got rid of the function that returns a function to strip the full path and rewrote it to take the full workspace path as a parameter. I thought this simplified the code a bit.
- Got rid of the global project root path.
- Changed the status bar to show when storyteller is not active (and made it clickable to start using it) and when it is active.