Made any initial code NOT relevant when a workspace is being tracked.
On all insert, create file, and create directory events generated from a reconciliation with some existing code these events are marked as not relevant so that they will not contribute to the active playback. All initial code will NOT be animated.

I also fixed a bug in placing the comment marker when adding a new comment.