Can see events past the last relevant one
Removed the feature of making the last event relevant so that one can see all of the code after the relevant events. If there is code after the last relevant event it can now be played by moving forward at the end of the playback.

Updated and improved the code that made it so that a project with existing code (before storyteller begins tracking changes) will be marked as permanently never relevant.

Added a relevant and total event counter so the user can see where they are in both sets of events.

Completely removed the compression code.