When a new file has some text in it the text is added to the system.
Occassionally, a tool will generate files that have some text in it. In the past the contents of new files were not added to the system. Now, when a file is created the file is opened to see if it has some text in it. If it does then the contents are added as events.