Fixed bug when user uses menu to create a new file
Removed the default parameters in several functions.

Made a change where file/dirs are checked to see if they exist before using an id from the function getIdFromFilePath() and getIdFromDirPath().

Fixed a bug where any existing code in a project folder when the project is started is marked as permanently not relevant.

Fixed a bug when the user opens a new file using the menu instead of adding a file with the file explorer view.