All pastes now have a pasted insert event id.
If a paste happens from a copy outside of the editor then I store the text 'other' as the pasted event id in the insert event.

I will use this to know when a paste from outside storyteller happens (if the paste is from a copy that happened in storyteller then I already store the insert event's id in the Insert event's pastedEventId).