Updated HttpServer to return playback data
This commit adds functionality for a playback web page to request the necessary data to animate a playback using the browser's 'fetch' function.

This is different than the current system that gets all data in one big object that is embedded in the playback page.

The new routes include:
- get 'project' data (title, description, and branchId)
- set 'project' data (title and description only)
- get all developers
- get all developer groups
- get the editor's view of the files and directories (including renamed, moved, and deleted elements)
- get all comments
- get events with routes to
 -- get all events
 -- get all events from a starting index
 -- get all events from a starting index plus a requested number of additional events
 -- get all events from a starting index up to but not including an ending index
 -- get all events from the beginning of a playback until the first event that has a comment attached to it
 -- get all events from one after the starting index until the next event that has a comment attached to it
(These last two can be used to retrieve events from comment to comment)

I also moved a series of web editor routes to a new function. We could use these routes to create our own Ace editor and not need Visual Studio Code.