Copy public directory into open storyteller project
Created a public directory in this repo that holds html, css, js, and media files that we want to make available by the storyteller web server.

The new /core/public/ directory will be copied into an open storyteller project's /.storyteller/public/ directory to serve the playback file and any necessary js files.

At startup the latest and greatest versions of the /core/public/ files are copied into the project. All that is needed while debugging is to restart the server.