Dealing with windows CRLF
Windows editors store newlines as two characters '\r' (aka carriage return (CR)) and '\n' (aka linefeed (LF)). A windows VS Code instance includes these in inserted text.

In the past I treated CR and LF as separate inserted characters and played them back individually. However, Ace has a problem with that. It treats these all of these as newlines:
\n
\r\n
\r

So, a separate \r followed by a \n creates extra lines in the playback. I have changed the back end to look for \r\n and to store that as a NEWLINE in an insert event. The \r is still there but it is combined with the \n in the INSERT event.

When deleteing newlines the editor will have two characters CR LF but the insert in the file will just be a single event. In the event manager I decrease the number of events to delete for every deleted newline to account for this mismatch.