Small bug on comment with highlights in a different file
I noticed that when there are two comments in the same comment group but they have highlights in different files that the active highlighted file is not coming to the screen.

This also happens if there is a single comment in the group but the highlighted text is in a file different than the last updated file.

In the createCommentCard() function in the click listener I simulated a link click if a different file is on the screen.