Dev images in comments and playback
This commit is mostly about adding developer image avatars. I changed the developer class to include a URL for gravatar images (gravatar is a service that serves avatar images for free).

I added some code in playback to show who the author is of code and who the author of a comment is. It is pretty ugly right now and can be improved.

I also changed the name of a route from /editComment to what I think is more clear, /playbackEditable.

This commit also has the beginning of downloading code from the file system view but I have not made much progress yet.