Updated slider to have comments line up
I updated some slider code so that comments line up with events on the slider (mostly changing off by one errors).

I did not really understand the clickable tick mark code so I was unable to get that to work (it also seems to be off by one). I commented out the calls to this function and will add a bug to the 'known bug' list.