Call write() more frequently and remove description.
Call write() when adding/updating/deleting playback data like comments, devs, etc. instead of just at the end when there is a graceful shutdown. This was causing problems if the extension ever crashed because some recent data might have been lost.

Also removed the description from the project as it is now the first comment.

Lastly, I made some updates to the tests for these features and for the recent change to newline handling.