View playbacks from file system in addition to web server
Changed the way playbacks are served. In the past they were always served by a web server running in the extension. I am changing it so that playbacks can be viewed without a web server simply from the file file system in addition to the extension.

Instead of asking for all of the playback data through a series of fetch calls from the server in playbackInit.js I am creating and serving a single dynamic javascript file, /js/loadPlayback.js, that loads the global playbackData object.

Next, I added the ability for the extension to create a zip file with the static js and html along with the current data wrapped inside loadPlaybackData.js. This zip can be unzipped and someone can watch a playback without requiring any software other than a web browser.