Changed the way to check if a question was attached to a comment.
I changed the way that the UI decides if there is a question attached to a comment or not because every comment including the first one will have a questionCommentData (st-ui.js).

I updated model classes to use same shorter id check.

Question for Omar: I was a little confused by the name getCommentQuestion() it seems like createAnswerInput() might be more clear, no? What about the other related function names?