Paste images on the clipboard into a new/editable comment.
This feature allows the user to copy an image to the clipboard (using the file explorer or a screenshot tool) and then paste it into the media pop up in a single step. This saves the user from having to save the screenshot and then select it from the file system.

Bug fixes: deleting audios and videos was not working I updated the query selector to fix it. I also moved hiding the cursor to the function that creates editors.