Removed Ace error markers and changed the order that media is added to the gallery.
Ace has a feature where it attempts to identify errors in code and give suggestions on how to fix it. I removed this.

Any new media that gets added was added at the end of the media container making it hard to see what you just added (when there is already a lot of media in the gallery). I changed it so that new media files are added at the top instead of the bottom of the gallery.