First attempt at an optimization to reduce the number of events that have to be adjusted.
I keep track of the position of the earliest removed event. All events that come after that one have to be adjusted. All the ones that come before it don't. This should reduce the expensive lineNumber, column, and previousNeighbor adjustments for events being handled in a 'perfect programmer' reduction of events.