Added the ability to rewrite events for 'perfect programmer' scenarios
An author can now choose to rewrite between events so that anything that was added and then removed between the events is filtered out.

There are two types of filtering:
- between all comments
- between two tags

The first cleans up all of the superflous events between each pair of comments. The edits happen in order of:
1. file system events
2. deletes
3. inserts
So that the viewer will have no real idea of how the code was actually written. It will appear as all the deletes happening first and then all of the inserts from left-to-right, top-to-bottom.

The second requires that the user add a start and end tag to at least two comments. The superflous events are removed and the code is played back in the order that the programmer wrote it.