Adds a context menu to search through selected text in the editor
This change adds a context menu to copy code and to start a selected text search.

If some code is selected and a search is performed then all of the comments will be checked to see if any of the selected text is part of a comment. If so, those comments are shown and the others are filtered out.

Also, fixed a bug where an extra event was being added to the comment's selected text event ids because of the way that the Ace range works.