Updated for text-to-speech reading of the comments.
A user can hear a comment be read to them. When creating a playback a user can upload an mp3 file with their narration. Alternatively, if the AI assistant is present then an AI generated audio file can be generated.

This works when viewing playbacks too. Either the uploaded file will play or an AI generated file can be built.