Added url queries to playbacks
The user can specify which comment number of event number that they would like to start the playback on.

When moving through a playback, the url is updated to reflect the active comment or the event number.

The url's will now look like this:
/index.html?comment=8
/index.html?event=152

Also, I removed the observability watcher in blog mode as it was interfering with this feature.